added useApi as a custom hook
